+++ Pandora/Paradigm/Controlflow/Effect/Transformer/Monadic.hs
@@ -2,7 +2,7 @@
 module Pandora.Paradigm.Controlflow.Effect.Transformer.Monadic (Monadic (..), (:>) (..)) where
 
 import Pandora.Core.Functor (type (<))
-import Pandora.Core.Interpreted (Schematic, Interpreted (Primary, run, unite, (<~~~)))
+import Pandora.Core.Interpreted (Interpreted (Primary, run, unite, (<~~~)))
 import Pandora.Pattern.Morphism.Straight (Straight (Straight))
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category ((<--), (<---), (<----))
@@ -21,6 +21,7 @@ import Pandora.Paradigm.Algebraic.Product ((:*:)((:*:)))
 import Pandora.Paradigm.Algebraic.Sum ((:+:))
 import Pandora.Paradigm.Algebraic.One (One (One))
 import Pandora.Paradigm.Algebraic (Pointable, point)
+import Pandora.Paradigm.Schemes (Schematic)
 
 class Interpreted m t => Monadic m t where
 	{-# MINIMAL wrap #-}