+++ Pandora/Paradigm/Inventory/Some/Store.hs
@@ -2,7 +2,7 @@
 module Pandora.Paradigm.Inventory.Some.Store where
 
 import Pandora.Core (type (:.), type (>), type (>>>), type (<:=), type (~>))
-import Pandora.Core.Interpreted (Interpreted (Primary, run, unite, (=#-)), Schematic)
+import Pandora.Core.Interpreted (Interpreted (Primary, run, unite, (=#-)))
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category ((<--), (<---), (<----), (<-----), (<------), identity)
 import Pandora.Pattern.Kernel (constant)
@@ -20,7 +20,7 @@ import Pandora.Pattern.Morphism.Flip (Flip (Flip))
 import Pandora.Pattern.Morphism.Straight (Straight (Straight))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
 import Pandora.Paradigm.Controlflow.Effect.Transformer.Comonadic (Comonadic (bring), (:<) (TC))
-import Pandora.Paradigm.Schemes.TUT (TUT (TUT), type (<:<.>:>))
+import Pandora.Paradigm.Schemes (Schematic, TUT (TUT), type (<:<.>:>))
 
 -- | Context based computation on value
 newtype Store s a = Store ((:*:) s :. (->) s >>> a)